// JcCar-client/src/app/shared/services/model/alertInvolved.ts
import { StatusApplication } from './statusApplication';
/**
 * ALERTA API
 * SMS - Módulo Alertas de SMS que tem como objetivo comunicar aos interessados deliberações das Agências Reguladoras
 * e eventos de segurança relevantes para que as medidas necessárias sejam tomadas
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ActionPlan } from './actionPlan';
import { Alert } from './alert';
import { AttachmentInvolved } from './attachmentInvolved';


export interface AlertInvolved {
    actionsPlans?: Array<ActionPlan>;
    alert?: Alert;
    attachmentsForNotApplicable?: Array<AttachmentInvolved>;
    company?: string;
    comprehensiveness?: boolean;
    id?: number;
    justificationForNotApplicable?: string;
    registeredDate?: string;
    statusInvolved?: StatusApplication;
    responsedDate?: string;
    userKey?: string;
    viewed?: boolean;
}